rag_federal_individual/requirements.txt
httpx>=0.27.0
pypdf>=4.0.0
beautifulsoup4>=4.12.0
openai>=1.40.0
numpy>=1.26.0
python-dotenv>=1.0.0
tqdm>=4.66.0
